prevent opening files that are already open
tempfile.NamedTemporaryFile holds the file open, but we have been reopening
it (using the filename to get a file pointer); this can cause problems on
some systems.

Changed SafeFile to return a python file object, to avoid reopening the file.
Added SafeFilename to return a filename that doesn't have an open fd. This is
necessary to support FITS files, as cfitsio doesn't have an interface that
accepts a FILE* or fd.