make bypass functions non-deferred
After doing a locate, if the dataset has a bypass function
butler can't know if the location will work or not (the
dataset may actaully exist in a parent registry, or not,
or may not be used at all). To know if a bypass function
is for a given repo/storage, the function must be called
before the deferral mechanism.