test: robustify TestParentRepository
The shutil.rmtree in the tearDown can fail (e.g., "Directory not empty"
due to NFS dot files hanging around), which causes the whole test to fail.
Allow it to fail. More important is that the directory not exist when the
test starts, or more severe failures can occur that can delude the user
into thinking something important is wrong.